Further optimized the snake game. Added replays of a simulation.
New snake game algorithm is O(1), meaning it is not dependent on the size of the snake anymore. Also added replays of a simulation so you don't have to run the program and generate the data just to see the (most likely) end results.

+++ snake-genetic-algorithm/snake-genetic-algorithm/snake.hpp
@@ -35,7 +35,7 @@ private:
     int currentFitness;
     POS currentFruit;
     int tickCount;
-    double totalTickCount;
+    int totalTickCount;
 public:
     deque<POS> snake;
     int getFitness(){ return (currentFitness*currentFitness)/totalTickCount;}
